+++ contracts/sale/FixedSupplyLotSale.sol
@@ -5,9 +5,8 @@ pragma solidity 0.6.8;
 import "@animoca/ethereum-contracts-erc20_base/contracts/token/ERC20/IERC20.sol";
 
 import "./Sale.sol";
-import "../payment/KyberAdapter.sol";
 
-abstract contract FixedSupplyLotSale is Sale, KyberAdapter {
+abstract contract FixedSupplyLotSale is Sale {
 
     // a Lot is a class of purchasable sale items.
     struct Lot {
@@ -48,14 +47,12 @@ abstract contract FixedSupplyLotSale is Sale, KyberAdapter {
 
     /**
      * @dev Constructor.
-     * @param kyberProxy Kyber network proxy contract.
      * @param payoutWallet_ Account to receive payout currency tokens from the Lot sales.
      * @param payoutToken_ Payout currency token contract address.
      * @param fungibleTokenId Inventory token id of the fungible tokens bundled in a Lot item.
      * @param inventoryContract Address of the inventory contract to use in the delivery of purchased Lot items.
      */
     constructor(
-        address kyberProxy,
         address payable payoutWallet_,
         IERC20 payoutToken_,
         uint256 fungibleTokenId,
@@ -65,7 +62,6 @@ abstract contract FixedSupplyLotSale is Sale, KyberAdapter {
             payoutWallet_,
             payoutToken_
         )
-        KyberAdapter(kyberProxy)
         internal
     {
         setFungibleTokenId(fungibleTokenId);
@@ -276,56 +272,6 @@ abstract contract FixedSupplyLotSale is Sale, KyberAdapter {
     }
 
     /**
-     * @dev Retrieves user purchase price information for the given quantity of Lot items.
-     * @param recipient The user for whom the price information is being retrieved for.
-     * @param lotId Lot id of the items from which the purchase price information will be retrieved.
-     * @param quantity Quantity of Lot items from which the purchase price information will be retrieved.
-     * @param tokenAddress Purchase currency token contract address.
-     * @return minConversionRate Minimum conversion rate from purchase tokens to payout tokens.
-     * @return totalPrice Total price (excluding any discounts), in purchase currency tokens.
-     * @return totalDiscounts Total discounts to apply to the total price, in purchase currency tokens.
-     */
-    function getPrice(
-        address payable recipient,
-        uint256 lotId,
-        uint256 quantity,
-        IERC20 tokenAddress
-    )
-        external
-        view
-        returns
-    (
-        uint256 minConversionRate,
-        uint256 totalPrice,
-        uint256 totalDiscounts
-    )
-    {
-        Lot memory lot = _lots[lotId];
-
-        require(lot.exists);
-        require(tokenAddress != IERC20(0));
-
-        (totalPrice, totalDiscounts) = _getPrice(recipient, lot, quantity);
-
-        if (tokenAddress == payoutToken) {
-            minConversionRate = 1000000000000000000;
-        } else {
-            (, uint tokenAmount) = _convertToken(payoutToken, totalPrice, tokenAddress);
-            (, minConversionRate) = kyber.getExpectedRate(tokenAddress, payoutToken, tokenAmount);
-
-            if (totalPrice > 0) {
-                totalPrice = ceilingDiv(totalPrice.mul(10**36), minConversionRate);
-                totalPrice = _fixTokenDecimals(payoutToken, tokenAddress, totalPrice, true);
-            }
-
-            if (totalDiscounts > 0) {
-                totalDiscounts = ceilingDiv(totalDiscounts.mul(10**36), minConversionRate);
-                totalDiscounts = _fixTokenDecimals(payoutToken, tokenAddress, totalDiscounts, true);
-            }
-        }
-    }
-
-    /**
      * Validates a purchase.
      * @param purchase Purchase conditions (extData[0]:max token amount,
      *  extData[1]:min conversion rate).
@@ -367,44 +313,6 @@ abstract contract FixedSupplyLotSale is Sale, KyberAdapter {
         priceInfo = new bytes32[](2);
         priceInfo[0] = bytes32(totalPrice);
         priceInfo[1] = bytes32(totalDiscounts);
-    }
-
-    /**
-     * Transfers the funds of a purchase payment from the purchaser to the
-     * payout wallet.
-     * @param purchase Purchase conditions (extData[0]:max token amount,
-     *  extData[1]:min conversion rate).
-     * @param priceInfo Implementation-specific calculated purchase price
-     *  information (0:total price, 1:total discounts).
-     * @return paymentInfo Implementation-specific purchase payment funds
-     *  transfer information (0:purchase tokens sent, 1:payout tokens received).
-     */
-    function _transferFunds(
-        Purchase memory purchase,
-        bytes32[] memory priceInfo
-    ) internal override virtual returns (bytes32[] memory paymentInfo) {
-        uint256 totalPrice = uint256(priceInfo[0]);
-        uint256 totalDiscounts = uint256(priceInfo[1]);
-        uint256 totalDiscountedPrice = totalPrice.sub(totalDiscounts);
-        uint256 maxTokenAmount = uint256(purchase.extData[0]);
-        uint256 minConversionRate = uint256(purchase.extData[1]);
-
-        (uint256 purchaseTokensSent, uint256 payoutTokensReceived) =
-            _swapTokenAndHandleChange(
-                purchase.paymentToken,
-                maxTokenAmount,
-                payoutToken,
-                totalDiscountedPrice,
-                minConversionRate,
-                purchase.operator,
-                address(uint160(address(this))));
-
-        require(payoutTokensReceived >= totalDiscountedPrice);
-        require(payoutToken.transfer(payoutWallet, payoutTokensReceived));
-
-        paymentInfo = new bytes32[](2);
-        paymentInfo[0] = bytes32(purchaseTokensSent);
-        paymentInfo[1] = bytes32(payoutTokensReceived);
     }
 
     /**